SimpleChecksums

Documentation for SimpleChecksums.

SimpleChecksums.additive_checksum

SimpleChecksums.bsd_checksum

SimpleChecksums.sysv_checksum

SimpleChecksums.additive_checksum — Method

additive_checksum(T, data, init::T = zero(T), modulo::T = typemax(T)) where {T<:Unsigned}

Compute a simple additive hash of the values in data.

Sums all the values in data, then computes modulo modulo, optionally starting from init instead of zero. Sums are allowed to overflow, which resets the sum to zero.

This function is extremely fast but suffers from many collisions. Flaws to keep in mind are:

zero values in data do not affect the checksum value at all;

the order of the values in data can be permuted without changing the checksum value; and

because of the modulo operation, the checksum of all zeros is equal to the checksum of all modulo values.

Arguments

T: An unsinged integer type.

data: A single Unsigned value or an iterator of values.

init::Unsigned = zero(T): Optional starting value.

modulo::T = typemax(T): A number to use as the modulo for addition. Typical choices are typemax(T) (the default) or some prime number close to but less than typemax(T).

Predefined convenience functions are additive16, additive32, and additive64.

See also: sysv16 for a different way of folding the sum into a fixed number of bits.

SimpleChecksums.bsd_checksum — Function

bsd_checksum(UInt16, data, init::UInt16 = zero(UInt16))

Compute a hash of data using the method defined in BSD's sum utility (and implemented in the GNU sum program).

The BSD checksum computes a 16-bit checksum value by bit-rotating the checksum value from the previous step to the right by 1 bit, then adding the next value from data and repeating. Sums are allowed to overflow, which resets the sum to zero.

This function is fast but suffers from many collisions. Flaws to keep in mind are:

zero values at the beginning of data do not affect the checksum value at all (unless the initial value is set to something other than zero); and

runs of zeros in data that are multiples of 16 in length do not affect the checksum value at all.

Arguments

data: A single Unsigned value or an iterator of values.

init::UInt16 = zero(UInt16): Optional starting value.

Predefined convenience function is bsd16.

SimpleChecksums.sysv_checksum — Function

sysv_checksum(UInt16, data, init::UInt16 = zero(UInt16))

Compute a hash of data using the method defined in SYS-V's sum utility (and implemented in the GNU sum program).

The SYS-V checksum computes a 16-bit checksum value using a simple 32-bit additive sum, then folding the sum into a 16-bit number by adding the low 16 bits to the high 16 bits (shifted right by 16 bits). This fold is performed twice to guarantee the result is a 16-bit number. Unlike the SYS-V and GNU implementations, the sum in this implementation is allowed to overflow, which resets the sum to zero (SYS-V and GNU will raise an error instead).

This function is very fast but suffers from many collisions. Flaws to keep in mind are:

zero values anywhere in data do not affect the checksum value at all (unless the initial value is set to something other than zero); and

the order of the data can be permuted without affecting the checksum value.

Arguments

data: A single Unsigned value or an iterator of values.

init::UInt16 = zero(UInt16): Optional starting value.

Predefined convenience function is sysv16.

See also: additive16.